+++ app/models/admin_user.py
@@ -1,3 +1,0 @@
-from .user import User as AdminUser
-
-__all__ = ["AdminUser"]